Hopefully fixed something in MoveDamage

+++ PokeHack/Pokemon.cs
@@ -113,9 +113,9 @@ namespace PokeHack
 		public int MoveDamage(Move move, Pokemon defender) {
             int damage = 0;
             if (String.Compare(move.DamageClass, "physical") == 0) 
-                damage = (int)((((float)(2 * this.Level + 10)/250 * (this.Attack /	defender.Defense) + 2) * GetModifier(move.Type, defender.Type1, defender.Type2)));
+                damage = (int)((((float)(2 * this.Level + 10)/250.0 * (((float) this.Attack) / defender.Defense) + 2) * GetModifier(move.Type, defender.Type1, defender.Type2)));
             else if(String.Compare(move.DamageClass, "special") == 0)
-                damage = (int)((((float)(2 * this.Level + 10) / 250 * (this.SpecialAttack / defender.SpecialDefense) + 2) * GetModifier(move.Type, defender.Type1, defender.Type2)));
+                damage = (int)((((float)(2 * this.Level + 10) / 250.0 * (((float) this.SpecialAttack) / defender.SpecialDefense) + 2) * GetModifier(move.Type, defender.Type1, defender.Type2)));
             Console.WriteLine(move.DamageClass + " " + damage);
             return damage;
 		}
